fix: update tooltip (#2037)
* fix: update tooltip

* fix: rebuild

* fix: failing build

* fix: speed-up

* fix: lock file

* fix: rebuild

* fix: rebuild

* fix: rebuild

* fix: rebuild

* fix: rebuild

* fix: update unit tests

* fix: update unit tests
