+++ workspaces/core/src/tooltip/private/types/Align.ts
@@ -1,0 +1,3 @@
+type Align = 'start' | 'center' | 'end' | undefined;
+
+export default Align;
