+++ workspaces/core/src/tooltip/private/types/ArrowProps.ts
@@ -1,0 +1,5 @@
+type ArrowProps = {
+  className?: string;
+};
+
+export default ArrowProps;
