+++ workspaces/core/src/tooltip/private/types/ContentProps.ts
@@ -1,0 +1,9 @@
+import type Position from './Position';
+
+type ContentProps = {
+  position?: Position;
+  className?: string;
+  children?: React.ReactNode;
+};
+
+export default ContentProps;
