+++ workspaces/core/src/tooltip/private/types/Position.ts
@@ -1,0 +1,11 @@
+type Position =
+  | 'top-start'
+  | 'top-center'
+  | 'top-end'
+  | 'bottom-start'
+  | 'bottom-center'
+  | 'bottom-end'
+  | 'right'
+  | 'left';
+
+export default Position;
