+++ workspaces/core/src/tooltip/private/types/Side.ts
@@ -1,0 +1,3 @@
+type Side = 'top' | 'right' | 'bottom' | 'left';
+
+export default Side;
