+++ workspaces/core/src/tooltip/private/types/TooltipRootProps.ts
@@ -1,0 +1,5 @@
+type TooltipRootProps = {
+  children?: React.ReactNode;
+};
+
+export default TooltipRootProps;
