+++ workspaces/core/src/tooltip/private/types/TooltipState.ts
@@ -1,0 +1,6 @@
+type TooltipState = {
+  withArrow?: boolean;
+  registerChild?: (child: string) => () => void;
+};
+
+export default TooltipState;
