+++ workspaces/core/src/tooltip/private/types/TriggerProps.ts
@@ -1,0 +1,6 @@
+type TriggerProps = {
+  className?: string;
+  children?: React.ReactNode;
+};
+
+export default TriggerProps;
