+++ workspaces/core/src/tooltip/private/utils/TooltipContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type TooltipState from '../types/TooltipState';
+
+const TooltipContext = createContext<TooltipState>({});
+TooltipContext.displayName = 'TooltipContext';
+
+export default TooltipContext;
